requirements.txt
dgl-cu101==0.6.1
torch==1.4.0
scikit-learn==1.0.2
numpy==1.19.2
networkx==2.5
absl-py==0.14.1
scipy==1.5.2